2020.2

+++ Machine_Learning/Introduction/03-Basic/Module_7/build_app_hls.sh
@@ -15,9 +15,35 @@
 #
 
 #!/bin/bash
+set -e
+BOARD_IP=10.176.18.178
 
-git pull
-rm -rf build
-mkdir build&&cd build
-cmake -DUSE_DRM=on -DUSE_KERNEL=on .. && make
-cp *drm ../
+build_command(){
+    cmake -DUSE_DRM=on -DUSE_KERNEL=on .. && make
+    cp *drm ../
+
+}
+
+clean_workspace(){
+    if [ -d build ]; then
+    rm -rf build
+    fi
+    mkdir -pv build
+    cd build
+    
+}
+copy_file()
+{
+    scp usb_input* root@$BOARD_IP:/home/root/
+    scp HLS_*/lib* root@$BOARD_IP:/usr/lib
+    scp HLS_*/test_hls_kernel root@$BOARD_IP:/home/root/
+}
+
+main(){
+    clean_workspace
+    git pull
+    build_command
+    copy_file
+}
+
+main @$
